gdb: use existing argument more in rename_vmcore_idle_reg_sections
In corelow.c, in the function rename_vmcore_idle_reg_sections, the
argument ABFD holds the core file bfd pointer.  When this function is
called current_program_space->core_bfd() is passed as the argument
value.

Within this function, we sometimes use the function argument, and
sometimes access current_program_space->core_bfd() directly.

This is confusing, and unnecessary.  Lets not do that.

I've renamed the argument to cbfd (for Core file BFD), and then
updated the function to make use of this argument throughout.  This
reduces the number of accesses to global state, which is, I think, a
good thing.

There should be no user visible changes after this commit.

Approved-By: Simon Marchi <simon.marchi@efficios.com>