libsframe: use offsets to FDE and FRE sub-sections
In SFrame format, the SFrame FDEs begin at an offset 'sfh_fdeoff' from
the end of the SFrame header.  Similarly, the SFrame FREs begin at an
offset 'sfh_freoff' from the end of the SFrame header.

While the GNU assembler generates these subsections such that sfd_fdeoff
is zero, the correct way to decode the SFrame section (conforming with
the specification) is to use both sfh_fdeoff and sfh_freoff in
sframe_decode.

libsframe/
	* sframe.c (sframe_decode): Use offsets to SFrame FDE and FRE
	sub-sections as applicable.